add confirm dialog box for deletion

+++ src/components/GroceryContextProvider.jsx
@@ -60,7 +60,19 @@ const GroceryContextProvider = ({children}) => {
             setSwalBg('#f8f9fa');
             setSwalColor('#212529');
         }
-    },[darkMode]);    
+    },[darkMode]);   
+    
+    const displaySwalFire = (swalTitle, swalText) => { //show alert message if task was updated/MarkedAsDone.
+        Swal.fire({
+            icon: 'success',
+            title: `${swalTitle}`,
+            text: `${swalText}`,
+            color: swalColor,
+            background: swalBg, 
+            showConfirmButton: false,
+            timer: 3000
+        });
+    }
 
     const submitGroceryItem = (event)=>{
         event.preventDefault(); //to prevent webpage from loading
@@ -80,35 +92,45 @@ const GroceryContextProvider = ({children}) => {
     };
 
     const removeGroceryItem = (groceryItemId, taskName) =>{ //delete specific item to an array
-        const filteredItems = groceryList.filter((items) => items.id !== groceryItemId);
-        setGroceryList(filteredItems);
-        localStorage.setItem('localGroceryList', JSON.stringify(filteredItems));
         Swal.fire({
-            title: 'Task deleted!',
-            text: `Task "${taskName}" was successfully deleted!`,
+            title: `Delete task "${taskName}"?`,
+            text: "You won't be able to revert this!",
+            icon: 'warning',
             color: swalColor,
             background: swalBg, 
-            iconHtml: '<i class="bi bi-trash3"></i>',
-            customClass:{
-                icon: `${Style.iconBorder}`,
-            },
-            showConfirmButton: false,
-            timer: 3000
-        });
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                const filteredItems = groceryList.filter((items) => items.id !== groceryItemId);
+                setGroceryList(filteredItems);
+                localStorage.setItem('localGroceryList', JSON.stringify(filteredItems));
+                
+                Swal.fire({
+                    title: 'Task deleted!',
+                    text: `Task "${taskName}" was successfully deleted!`,
+                    color: swalColor,
+                    background: swalBg, 
+                    iconHtml: '<i class="bi bi-trash3"></i>',
+                    customClass:{
+                        icon: `${Style.iconBorder}`,
+                    },
+                    showConfirmButton: false,
+                    timer: 3000
+                });
+            }
+          })
     };
 
     const markAsDone = (taskId, taskName) => { //mark task item as done
         setTaskDone([{id: taskId, name: taskName.trim()}, ...taskDone]);
         removeGroceryItem(taskId);
-        Swal.fire({
-            icon: 'success',
-            title: 'Good Job!',
-            text: `Task "${taskName}" was successfully completed!`,
-            color: swalColor,
-            background: swalBg,             
-            showConfirmButton: false,
-            timer: 3000
-        });
+
+        let swalText = `Task "${taskName}" was successfully completed!`;
+
+        displaySwalFire('Good Job!', swalText); //display alert message
     };
 
     const editTaskItem = (taskId, taskName) => { //edit task item   
@@ -125,15 +147,9 @@ const GroceryContextProvider = ({children}) => {
         setGroceryList(updateTaskItem);  
 
         if (updateIndicator) {
-            Swal.fire({
-                icon: 'success',
-                title: 'Task Updated!',
-                text: `Task was updated to "${taskName}".`,
-                color: swalColor,
-                background: swalBg, 
-                showConfirmButton: false,
-                timer: 3000
-            });    
+            let swalText = `Task was updated to "${taskName}".`;
+
+            displaySwalFire('Task Updated!', swalText); //display alert message
         }
     };
